AIDD-30-DAY-Task4/requirements.txt
streamlit
pypdf2
google-generativeai
python-dotenv
